requirements.txt
streamlit
pandas
plotly
scikit-learn
numpy
fpdf2
